requirements.txt
asgiref
Django
django-js-asset
django-mptt
djangorestframework
dnspython
pymongo
python-dotenv
sqlparse
typing_extensions
